main페이지 따로 분리 및 main.js에서는 맨 처음 렌더 시  renderRoute가 동작되게 함
